refactor(cli): modularize init.rs into focused submodules
Split the ~1900-line init.rs into a thin facade plus six cohesive
submodules under init/: provider metadata, model listing, pure
ranking, the interactive prompter, spec resolution, and rendering.
Tests are co-located with the code they exercise, with shared
fixtures hoisted into a cfg(test) test_support module.

No behavior change. The public surface (run_init, InitArgs) is
unchanged; everything else is tightened to pub(crate) within the
subtree.

Ref: LOG-000000 (no ref)
