+++ crates/aura-cli/src/init/render.rs
@@ -1,0 +1,222 @@
+//! Rendering: turn a resolved `ConfigSpec` into the generated `config.toml`
+//! and (when a key was typed) a `.env`, plus the human-readable next-steps
+//! shown afterwards.
+
+use std::path::Path;
+
+#[cfg(feature = "standalone-cli")]
+use anyhow::Result;
+
+use super::spec::ConfigSpec;
+
+fn toml_escape(s: &str) -> String {
+    s.replace('\\', "\\\\").replace('"', "\\\"")
+}
+
+/// Render the `[agent.llm]` block. Values that come from the environment are
+/// referenced via `{{ env.<ACTUAL_VAR> }}` — no intermediate `LLM_*` rename.
+pub(crate) fn render_config(spec: &ConfigSpec) -> String {
+    let mut llm = format!("provider = \"{}\"\n", spec.provider.as_str());
+    if let Some(var) = spec.api_key_env_var() {
+        llm.push_str(&format!("api_key = \"{{{{ env.{var} }}}}\"\n"));
+    }
+    llm.push_str(&format!("model = \"{}\"\n", toml_escape(&spec.model)));
+    if let Some(region) = &spec.region {
+        llm.push_str(&format!("region = \"{}\"\n", toml_escape(region)));
+    }
+    if let Some(base_url) = &spec.base_url {
+        llm.push_str(&format!("base_url = \"{}\"\n", toml_escape(base_url)));
+    }
+
+    format!(
+        "# Generated by `aura-cli init`.\n\
+         \n\
+         [agent]\n\
+         name = \"{name}\"\n\
+         system_prompt = \"You are a helpful general-purpose assistant.\"\n\
+         \n\
+         [agent.llm]\n\
+         {llm}",
+        name = toml_escape(&spec.name),
+    )
+}
+
+/// Render a `.env` file for a user-provided API key. Only called when the
+/// user typed a key that isn't already in their environment.
+pub(crate) fn render_env(env_var: &str, value: &str) -> String {
+    format!(
+        "# Generated by `aura-cli init`. Contains a secret — do not commit it \
+         (add this file to your .gitignore).\n\
+         {env_var}={value}\n"
+    )
+}
+
+/// Merge a new key into an existing `.env`: replace the line if the key
+/// already exists, otherwise append it.
+pub(crate) fn merge_env(existing: &str, env_var: &str, value: &str) -> String {
+    let mut found = false;
+    let mut out = String::new();
+    for line in existing.lines() {
+        let trimmed = line.trim_start();
+        if trimmed
+            .strip_prefix(env_var)
+            .is_some_and(|rest| rest.trim_start().starts_with('='))
+        {
+            out.push_str(&format!("{env_var}={value}\n"));
+            found = true;
+        } else {
+            out.push_str(line);
+            out.push('\n');
+        }
+    }
+    if !found {
+        out.push_str(&format!("{env_var}={value}\n"));
+    }
+    out
+}
+
+/// Validate the generated config through the real parser.
+#[cfg(feature = "standalone-cli")]
+pub(crate) fn validate_rendered(spec: &ConfigSpec, rendered: &str) -> Result<()> {
+    let mut literal = rendered.to_string();
+    // The only env reference in a generated config is the api_key; substitute a
+    // literal so the real parser can validate the rest (provider/model are
+    // already literals).
+    if let Some(var) = spec.api_key_env_var() {
+        literal = literal.replace(&format!("{{{{ env.{var} }}}}"), "test-key");
+    }
+    aura_config::load_config_from_str(&literal)
+        .map_err(|e| anyhow::anyhow!("generated config failed validation (bug): {e}"))?;
+    Ok(())
+}
+
+/// Human-readable next-steps shown after writing the files.
+pub(crate) fn next_steps(config_path: &Path, wrote_env: bool) -> String {
+    let dir = config_path.parent().filter(|p| !p.as_os_str().is_empty());
+    let (run_prefix, config_for_run) = match dir {
+        Some(d) => {
+            let name = config_path.file_name().map_or_else(
+                || config_path.display().to_string(),
+                |n| n.to_string_lossy().into_owned(),
+            );
+            (format!("cd {} && ", d.display()), name)
+        }
+        None => (String::new(), config_path.display().to_string()),
+    };
+    let env_note = if wrote_env {
+        "\nThe API key was written to .env (gitignored — do not commit it).\n\
+         The server reads it automatically; a shell export of the same \
+         variable takes precedence."
+    } else {
+        ""
+    };
+    format!(
+        "\nNext steps:\n  \
+           1. {run_prefix}CONFIG_PATH={config_for_run} aura-web-server\n  \
+           2. aura-cli --api-url http://localhost:8080\n\
+         {env_note}",
+    )
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::init::provider::{DEFAULT_OLLAMA_URL, Provider};
+    use crate::init::test_support::{args, resolve};
+
+    #[test]
+    fn deterministic_output() {
+        let spec1 = resolve(&args()).unwrap();
+        let spec2 = resolve(&args()).unwrap();
+        assert_eq!(render_config(&spec1), render_config(&spec2));
+    }
+
+    #[test]
+    fn openai_config_references_native_env_var() {
+        let spec = resolve(&args()).unwrap();
+        let rendered = render_config(&spec);
+        assert!(rendered.contains("provider = \"openai\""));
+        assert!(rendered.contains("api_key = \"{{ env.OPENAI_API_KEY }}\""));
+        assert!(rendered.contains("model = \"gpt-5.1\""));
+        assert!(rendered.contains("name = \"assistant\""));
+        toml::from_str::<toml::Value>(&rendered).unwrap();
+    }
+
+    #[test]
+    fn bedrock_gets_region_and_no_key() {
+        let mut a = args();
+        a.provider = Some(Provider::Bedrock);
+        a.region = Some("us-east-1".to_string());
+        let spec = resolve(&a).unwrap();
+        let rendered = render_config(&spec);
+        assert!(rendered.contains("region = \"us-east-1\""));
+        assert!(!rendered.contains("api_key"));
+    }
+
+    #[test]
+    fn ollama_gets_base_url_and_no_key() {
+        let mut a = args();
+        a.provider = Some(Provider::Ollama);
+        let spec = resolve(&a).unwrap();
+        let rendered = render_config(&spec);
+        assert!(rendered.contains(&format!("base_url = \"{DEFAULT_OLLAMA_URL}\"")));
+        assert!(!rendered.contains("api_key"));
+    }
+
+    #[test]
+    fn render_env_single_key() {
+        let env = render_env("OPENAI_API_KEY", "sk-xyz");
+        assert!(env.contains("OPENAI_API_KEY=sk-xyz"));
+        assert!(!env.contains("LLM_"));
+    }
+
+    #[test]
+    fn merge_env_preserves_unrelated() {
+        let existing = "GITHUB_TOKEN=ghp_abc\nOPENAI_API_KEY=old\n";
+        let env = merge_env(existing, "OPENAI_API_KEY", "sk-new");
+        assert!(env.contains("GITHUB_TOKEN=ghp_abc"));
+        assert!(env.contains("OPENAI_API_KEY=sk-new"));
+        assert!(!env.contains("OPENAI_API_KEY=old"));
+    }
+
+    #[test]
+    fn merge_env_appends_when_absent() {
+        let existing = "GITHUB_TOKEN=ghp_abc\n";
+        let env = merge_env(existing, "OPENAI_API_KEY", "sk-new");
+        assert!(env.contains("GITHUB_TOKEN=ghp_abc"));
+        assert!(env.contains("OPENAI_API_KEY=sk-new"));
+    }
+
+    #[test]
+    fn merge_env_idempotent() {
+        let once = merge_env("", "OPENAI_API_KEY", "sk-test");
+        assert_eq!(merge_env(&once, "OPENAI_API_KEY", "sk-test"), once);
+    }
+
+    #[test]
+    fn next_steps_mentions_config() {
+        let s = next_steps(Path::new("config.toml"), false);
+        assert!(s.contains("config.toml"), "got: {s}");
+    }
+
+    #[test]
+    fn next_steps_mentions_env_when_written() {
+        let s = next_steps(Path::new("config.toml"), true);
+        assert!(s.contains(".env"), "got: {s}");
+    }
+
+    #[test]
+    fn next_steps_cds_into_config_dir() {
+        let s = next_steps(Path::new("proj/config.toml"), false);
+        assert!(s.contains("cd proj"), "got: {s}");
+        assert!(s.contains("CONFIG_PATH=config.toml"), "got: {s}");
+    }
+
+    #[cfg(feature = "standalone-cli")]
+    #[test]
+    fn validate_rendered_accepts_generated() {
+        let spec = resolve(&args()).unwrap();
+        let rendered = render_config(&spec);
+        validate_rendered(&spec, &rendered).unwrap();
+    }
+}
